asset property widget file restructured
