+++ asset-properties-widget/asset-properties.module.ts
@@ -3,19 +3,19 @@ import { CoreModule, HOOK_COMPONENTS, RealtimeModule } from '@c8y/ngx-components
 import * as preview from './common/preview';
 import { AssetPropertiesViewComponent } from './component/asset-properties-view/asset-properties-view.component';
 import { AssetPropertiesConfigComponent } from './component/asset-properties-config/asset-properties.config.component';
-import { AssetPropertiesSelectorComponent } from './component/asset-properties-config/asset-property-selector.component';
 import { AssetSelectorModule } from '@c8y/ngx-components/assets-navigator';
-import { schemaPropertySelectorCtrl } from './component/asset-property-selector-model/schema-property-selector.component';
-import { AssetPropertiesComponent } from './asset-properties-widget-view/asset-properties.component';
-import { AssetPropertiesItemComponent } from './asset-properties-widget-view/asset-properties-item.component';
 import { DragDropModule } from '@angular/cdk/drag-drop';
 import { Ng2SearchPipeModule } from 'ng2-search-filter';
+import { AssetPropertiesComponent } from './component/asset-properties-view/asset-properties/asset-properties.component';
+import { AssetPropertiesItemComponent } from './component/asset-properties-view/asset-properties/asset-properties-item.component';
+import { assetPropertyItemSelectorCtrl } from './component/asset-properties-config/asset-property-item-selector/asset-property-item-selector.component';
+import { AssetPropertiesSelectorComponent } from './component/asset-properties-config/asset-property-selector/asset-property-selector.component';
 
 @NgModule({
   declarations: [
     AssetPropertiesConfigComponent,
     AssetPropertiesSelectorComponent,
-    schemaPropertySelectorCtrl,
+    assetPropertyItemSelectorCtrl,
     AssetPropertiesViewComponent,
     AssetPropertiesComponent,
     AssetPropertiesItemComponent
@@ -35,7 +35,7 @@ import { Ng2SearchPipeModule } from 'ng2-search-filter';
         id: 'asset-properties-widget',
         label: 'Asset Properties',
         previewImage: preview.image,
-        description: 'Editable form for asset properties',
+        description: 'Editable form for asset properties widget',
         component: AssetPropertiesViewComponent,
         configComponent: AssetPropertiesConfigComponent,
         data: {
