+++ asset-properties-widget/component/asset-properties-config/asset-property-item-selector/asset-property-item-selector.component.html
