+++ asset-properties-widget/component/asset-properties-config/asset-property-selector/asset-property-selector.component.html
