+++ asset-properties-widget/component/asset-properties-view/asset-properties/asset-properties.model.ts
